Added a Discard function as well as some Helper functions
Use discard to attempt to remove a card. If can be legally discarded,
it's removed.
add() will add a given card to the bottom of a given column
remove() will remove the bottom card from a given column, if possible
getTop() will return the card at the top of the given column, if
possible, if the column is empty, it returns null